requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
google-cloud-datastore==1.7.3
